Improve setting pixel classification overlays
Moved PixelClassificationOverlay to the main GUI module for easier reuse.

Also moved set/getPixelLayer from PixelClassificationImageServer to ObservableMeasurementTableData, and it is now called only from QuPathViewer. This ensures live measurements should remain in sync with whatever the custom pixel classification overlay is showing.
